physical/lock/base/client/requirements.txt
paho-mqtt==1.2
pyserial==3.2.1
rpi.gpio==0.6.3
